requirements.txt
# Core RAG Framework
langchain==1.2.0
langchain-openai==1.1.5
langchain-community==0.4.1

# LLM Providers
openai==2.13.0
anthropic==0.75.0

# Vector Databases
pinecone==8.0.0
faiss-cpu==1.13.1

# Embeddings & Reranking
sentence-transformers==5.2.0
cohere==5.20.0

# Search
rank-bm25==0.2.2

# API & Web
fastapi==0.125.0
uvicorn[standard]==0.38.0
pydantic==2.12.5
python-multipart==0.0.21

# UI
streamlit==1.52.2

# Utilities
python-dotenv==1.2.1
tenacity==9.1.2
pandas==2.3.3
numpy==2.3.5

# Monitoring & Observability
prometheus-client==0.23.1
langsmith==0.5.0

# Development
pytest==9.0.2
black==25.12.0
flake8==7.3.0
ipython==9.8.0
jupyter==1.1.0